+++ deployments/chef/test/integration/with_custom_preload_dotnet_instrumentation/test.rb
@@ -48,6 +48,9 @@ describe file('/etc/splunk/zeroconfig/dotnet.conf') do
   its('content') { should match /^SPLUNK_PROFILER_MEMORY_ENABLED=true$/ }
   its('content') { should match /^SPLUNK_METRICS_ENABLED=true$/ }
   its('content') { should match /^OTEL_EXPORTER_OTLP_ENDPOINT=#{otlp_endpoint}$/ }
+  its('content') { should match /^OTEL_EXPORTER_OTLP_PROTOCOL=grpc$/ }
+  its('content') { should match /^OTEL_METRICS_EXPORTER=none$/ }
+  its('content') { should match /^OTEL_LOGS_EXPORTER=none$/ }
 end
 
 describe service('splunk-otel-collector') do