+++ deployments/chef/test/integration/with_default_preload_instrumentation/test.rb
@@ -2,7 +2,6 @@ libsplunk_path = '/usr/lib/splunk-instrumentation/libsplunk.so'
 java_tool_options = '-javaagent:/usr/lib/splunk-instrumentation/splunk-otel-javaagent.jar'
 node_options = '-r /usr/lib/splunk-instrumentation/splunk-otel-js/node_modules/@splunk/otel/instrument'
 resource_attributes = 'splunk.zc.method=splunk-otel-auto-instrumentation-\d+\.\d+\.\d+'
-otlp_endpoint = 'http://127.0.0.1:4317'
 dotnet_home = '/usr/lib/splunk-instrumentation/splunk-otel-dotnet'
 
 describe package('splunk-otel-auto-instrumentation') do
@@ -32,7 +31,9 @@ describe file('/etc/splunk/zeroconfig/java.conf') do
   its('content') { should match /^SPLUNK_PROFILER_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_PROFILER_MEMORY_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_METRICS_ENABLED=false$/ }
-  its('content') { should match /^OTEL_EXPORTER_OTLP_ENDPOINT=#{otlp_endpoint}$/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_ENDPOINT.*/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_PROTOCOL.*/ }
+  its('content') { should_not match /.*OTEL_METRICS_EXPORTER.*/ }
 end
 
 describe file('/etc/splunk/zeroconfig/node.conf') do
@@ -42,7 +43,9 @@ describe file('/etc/splunk/zeroconfig/node.conf') do
   its('content') { should match /^SPLUNK_PROFILER_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_PROFILER_MEMORY_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_METRICS_ENABLED=false$/ }
-  its('content') { should match /^OTEL_EXPORTER_OTLP_ENDPOINT=#{otlp_endpoint}$/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_ENDPOINT.*/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_PROTOCOL.*/ }
+  its('content') { should_not match /.*OTEL_METRICS_EXPORTER.*/ }
 end
 
 describe file('/etc/splunk/zeroconfig/dotnet.conf') do
@@ -59,7 +62,10 @@ describe file('/etc/splunk/zeroconfig/dotnet.conf') do
   its('content') { should match /^SPLUNK_PROFILER_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_PROFILER_MEMORY_ENABLED=false$/ }
   its('content') { should match /^SPLUNK_METRICS_ENABLED=false$/ }
-  its('content') { should match /^OTEL_EXPORTER_OTLP_ENDPOINT=#{otlp_endpoint}$/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_ENDPOINT.*/ }
+  its('content') { should_not match /.*OTEL_EXPORTER_OTLP_PROTOCOL.*/ }
+  its('content') { should_not match /.*OTEL_METRICS_EXPORTER.*/ }
+  its('content') { should_not match /.*OTEL_LOGS_EXPORTER.*/ }
 end
 
 describe service('splunk-otel-collector') do