+++ internal/receiver/discoveryreceiver/endpoint_tracker_test.go
@@ -715,13 +715,16 @@ func TestEntityEmittingLifecycle(t *testing.T) {
 	require.NoError(t, err)
 	require.Zero(t, failed)
 	expectedLogs := expectedEvents.ConvertAndMoveToLogs()
+	waitChan := make(chan struct{})
 	require.EventuallyWithT(t, func(c *assert.CollectT) {
 		logs := <-ch
-		assert.NoError(c, plogtest.CompareLogs(expectedLogs, logs, plogtest.IgnoreTimestamp()))
+		if assert.NoError(c, plogtest.CompareLogs(expectedLogs, logs, plogtest.IgnoreTimestamp())) {
+			close(waitChan)
+		}
 	}, 1*time.Second, 50*time.Millisecond)
 
 	// Ensure that entities are not emitted anymore
-	time.Sleep(60 * time.Millisecond)
+	<-waitChan
 	assert.Empty(t, ch)
 }