+++ internal/signalfx-agent/pkg/monitors/memory/memory_linux.go
@@ -12,16 +12,16 @@ func (m *Monitor) makeMemoryDatapoints(memInfo *mem.VirtualMemoryStat, swapInfo 
 
 	return []*datapoint.Datapoint{
 		datapoint.New("memory.utilization", dimensions, datapoint.NewFloatValue(float64(used)/float64(memInfo.Total)*100), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.used", dimensions, datapoint.NewIntValue(int64(used)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.buffered", dimensions, datapoint.NewIntValue(int64(memInfo.Buffers)), datapoint.Gauge, time.Time{}),
+		datapoint.New("memory.used", dimensions, datapoint.NewIntValue(int64(used)), datapoint.Gauge, time.Time{}),                //nolint:gosec
+		datapoint.New("memory.buffered", dimensions, datapoint.NewIntValue(int64(memInfo.Buffers)), datapoint.Gauge, time.Time{}), //nolint:gosec
 		// for some reason gopsutil decided to add slab_reclaimable to cached which collectd does not
-		datapoint.New("memory.cached", dimensions, datapoint.NewIntValue(int64(memInfo.Cached-memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.slab_recl", dimensions, datapoint.NewIntValue(int64(memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.slab_unrecl", dimensions, datapoint.NewIntValue(int64(memInfo.Slab-memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.free", dimensions, datapoint.NewIntValue(int64(memInfo.Free)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.total", dimensions, datapoint.NewIntValue(int64(memInfo.Total)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.swap_total", dimensions, datapoint.NewIntValue(int64(swapInfo.Total)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.swap_free", dimensions, datapoint.NewIntValue(int64(swapInfo.Free)), datapoint.Gauge, time.Time{}),
-		datapoint.New("memory.swap_used", dimensions, datapoint.NewIntValue(int64(swapInfo.Used)), datapoint.Gauge, time.Time{}),
+		datapoint.New("memory.cached", dimensions, datapoint.NewIntValue(int64(memInfo.Cached-memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}),    //nolint:gosec
+		datapoint.New("memory.slab_recl", dimensions, datapoint.NewIntValue(int64(memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}),                //nolint:gosec
+		datapoint.New("memory.slab_unrecl", dimensions, datapoint.NewIntValue(int64(memInfo.Slab-memInfo.Sreclaimable)), datapoint.Gauge, time.Time{}), //nolint:gosec
+		datapoint.New("memory.free", dimensions, datapoint.NewIntValue(int64(memInfo.Free)), datapoint.Gauge, time.Time{}),                             //nolint:gosec
+		datapoint.New("memory.total", dimensions, datapoint.NewIntValue(int64(memInfo.Total)), datapoint.Gauge, time.Time{}),                           //nolint:gosec
+		datapoint.New("memory.swap_total", dimensions, datapoint.NewIntValue(int64(swapInfo.Total)), datapoint.Gauge, time.Time{}),                     //nolint:gosec
+		datapoint.New("memory.swap_free", dimensions, datapoint.NewIntValue(int64(swapInfo.Free)), datapoint.Gauge, time.Time{}),                       //nolint:gosec
+		datapoint.New("memory.swap_used", dimensions, datapoint.NewIntValue(int64(swapInfo.Used)), datapoint.Gauge, time.Time{}),                       //nolint:gosec
 	}
 }